abstract and no-raise monitor error
fixes #522
